Add files via upload
1. What Is Junit:
JUnit is a unit testing open-source framework for the Java programming language. Java Developers use this framework to write and execute automated tests. In Java, there are test cases that have to be re-executed every time a new code is added. This is done to make sure that nothing in the code is broken.

2. What is unit test:
Unit testing is the process of testing small isolated portions of a software application called units.

3. What is the diff between junit and unit test:
JUnit is a derivative part of Unittest.
Key benefits of Test:Unit.